Changed logic in model, edited view
Made GUI changes to look better and to show the information in the GUI instead of the console. Fixed the model's channel handling. No need for DataFrame, stored Audio Segment object as a part of the model instance, and created a new attribute for the raw data so we can use it as a numpy array.